LPS-107777 | LPS-110480 Simplify

+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-service/src/main/java/com/liferay/portal/workflow/metrics/internal/background/task/WorkflowMetricsReindexBackgroundTaskStatusMessageTranslator.java
@@ -116,7 +116,7 @@ public class WorkflowMetricsReindexBackgroundTaskStatusMessageTranslator
 		double indexerPercentage = count / (double)total;
 
 		double totalPercentage =
-			(indexerCount + indexerPercentage) / (double)indexerTotal;
+			(indexerCount + indexerPercentage) / indexerTotal;
 
 		return (int)Math.min(Math.ceil(totalPercentage * 100), 100);
 	}
